+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added kql.Builder struct for safe building of KQL statements from variables without use of 'Unsafe' mode.
 * Simpler handling of query parameters using kql.Parameters struct.
+### Security
+* No redirects are allowed by default.
 ### Fixed
 * Replace non-ascii characters in headers to be in line with the service.